requirements.txt
PyYAML
keyboard
vosk
pyttsx3
SpeechRecognition